(apro) common.mk: Add TAP-based 'check' infrastructure
The design of this is based on me reading the Automake manual section on
testing, and playing a bit with how it works.

This adjusts _go-common.mk to use patter[1] to translate `go test -v` to
TAP.  I used it because I knew of it, and I could get it working quickly.
I don't like things about its TAP output.  I don't like that it uses
`go test -v` instead of `go test -json`.  I would like to write a
`gotest2tap.go` that does better.

[1]: https://github.com/apg/patter