Go through the git history to put docs in files instead of commit messages.
I Looked for all commits that had a >1 line commit message, and tried to
put important bits in the appropriate files.  A lot of it ended up going
in to `CHANGELOG.md`.

I also scanned through the full git log for things that seemed to warrant
a CHANGELOG entry but didn't get a >1 line commit message.